@@ -1,20 +1,32 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ToastrModule } from 'ngx-toastr';
 import { AppComponent } from './app.component';
+import { FooterComponent } from './footer/footer.component';
+import { HeaderComponent } from './header/header.component';
+import { RegistrationModule } from './registration/registration.module';
+import { CaptchaModule } from './shared/components/captcha/captcha.module';
+import { ImportantBookingInfoModule } from './shared/components/important-booking-info/important-booking-info.module';
+import { TableTemplateModule } from './shared/components/table-template/table-template.module';
+import { ConfigService } from './shared/services/config.service';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [
         RouterTestingModule,
+        RegistrationModule,
+        HttpClientTestingModule,
+        CaptchaModule,
+        ImportantBookingInfoModule,
+        TableTemplateModule,
         ToastrModule.forRoot()
       ],
       declarations: [
-        AppComponent
+        AppComponent, HeaderComponent, FooterComponent
       ],
-      providers: [
-      ]
+      providers: [ ConfigService ]
     }).compileComponents();
   });
 
