@@ -2,6 +2,8 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ConfigService } from '../shared/services/config.service';
+import { GuidelinesComponent } from './guidelines/guidelines.component';
+import { ParkDetailsComponent } from './park-details/park-details.component';
 
 import { RegistrationComponent } from './registration.component';
 
@@ -11,7 +13,7 @@ describe('RegistrationComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [RegistrationComponent],
+      declarations: [RegistrationComponent, ParkDetailsComponent, GuidelinesComponent],
       imports: [
         RouterTestingModule,
         HttpClientTestingModule
