@@ -1,5 +1,6 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed, tick } from '@angular/core/testing';
+import { FormsModule } from '@angular/forms';
 import { ConfigService } from '../../services/config.service';
 import { CaptchaDataService } from './captcha-data.service';
 
@@ -15,7 +16,7 @@ describe('CaptchaComponent', () => {
 
     await TestBed.configureTestingModule({
       declarations: [CaptchaComponent],
-      imports: [HttpClientTestingModule],
+      imports: [HttpClientTestingModule, FormsModule],
       providers: [ConfigService, { provide: CaptchaDataService, useValue: captchaServiceSpy }]
     }).compileComponents();
   });
