@@ -1,4 +1,5 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { TableTemplateModule } from '../table-template/table-template.module';
 
 import { ListComponent } from './list.component';
 
@@ -8,7 +9,8 @@ describe('ListComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ListComponent]
+      declarations: [ListComponent],
+      imports: [ TableTemplateModule ]
     })
       .compileComponents();
   }));
